requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
google-genai>=1.0.0
openai>=1.40.0
